chore: use RAF + manual loop
