+++ packages/ui/client/auto-imports.d.ts
@@ -45,16 +45,10 @@ declare global {
   const disableCoverage: typeof import('./composables/navigation')['disableCoverage']
   const eagerComputed: typeof import('@vueuse/core')['eagerComputed']
   const effectScope: typeof import('vue')['effectScope']
+  const endRun: typeof import('./composables/summary')['endRun']
   const extendRef: typeof import('@vueuse/core')['extendRef']
-  const files: typeof import('./composables/tests-status')['files']
-  const filesFailed: typeof import('./composables/summary')['filesFailed']
-  const filesIgnore: typeof import('./composables/summary')['filesIgnore']
-  const filesRunning: typeof import('./composables/summary')['filesRunning']
-  const filesSkipped: typeof import('./composables/summary')['filesSkipped']
-  const filesSnapshotFailed: typeof import('./composables/summary')['filesSnapshotFailed']
-  const filesSuccess: typeof import('./composables/summary')['filesSuccess']
-  const filesTodo: typeof import('./composables/summary')['filesTodo']
-  const finished: typeof import('./composables/tests-status')['finished']
+  const files: typeof import('./composables/summary')['files']
+  const finished: typeof import('./composables/summary')['finished']
   const getCurrentBrowserIframe: typeof import('./composables/api')['getCurrentBrowserIframe']
   const getCurrentInstance: typeof import('vue')['getCurrentInstance']
   const getCurrentScope: typeof import('vue')['getCurrentScope']
@@ -126,26 +120,20 @@ declare global {
   const shouldOpenInEditor: typeof import('./composables/error')['shouldOpenInEditor']
   const showCoverage: typeof import('./composables/navigation')['showCoverage']
   const showDashboard: typeof import('./composables/navigation')['showDashboard']
+  const startRun: typeof import('./composables/summary')['startRun']
   const syncRef: typeof import('@vueuse/core')['syncRef']
   const syncRefs: typeof import('@vueuse/core')['syncRefs']
   const templateRef: typeof import('@vueuse/core')['templateRef']
-  const testStatus: typeof import('./composables/tests-status')['testStatus']
-  const tests: typeof import('./composables/summary')['tests']
-  const testsFailed: typeof import('./composables/summary')['testsFailed']
-  const testsIgnore: typeof import('./composables/summary')['testsIgnore']
-  const testsSkipped: typeof import('./composables/summary')['testsSkipped']
-  const testsSuccess: typeof import('./composables/summary')['testsSuccess']
-  const testsTodo: typeof import('./composables/summary')['testsTodo']
+  const testRunState: typeof import('./composables/tests-status')['testRunState']
+  const testStatus: typeof import('./composables/summary')['testStatus']
   const throttledRef: typeof import('@vueuse/core')['throttledRef']
   const throttledWatch: typeof import('@vueuse/core')['throttledWatch']
-  const time: typeof import('./composables/summary')['time']
   const toRaw: typeof import('vue')['toRaw']
   const toReactive: typeof import('@vueuse/core')['toReactive']
   const toRef: typeof import('vue')['toRef']
   const toRefs: typeof import('vue')['toRefs']
   const toValue: typeof import('vue')['toValue']
   const toggleDark: typeof import('./composables/dark')['toggleDark']
-  const totalTests: typeof import('./composables/summary')['totalTests']
   const triggerRef: typeof import('vue')['triggerRef']
   const tryOnBeforeMount: typeof import('@vueuse/core')['tryOnBeforeMount']
   const tryOnBeforeUnmount: typeof import('@vueuse/core')['tryOnBeforeUnmount']
