+++ packages/ui/client/components.d.ts
@@ -21,9 +21,6 @@ declare module 'vue' {
     Modal: typeof import('./components/Modal.vue')['default']
     ModuleTransformResultView: typeof import('./components/ModuleTransformResultView.vue')['default']
     Navigation: typeof import('./components/Navigation.vue')['default']
-    NewNavigation: typeof import('./components/NewNavigation.vue')['default']
-    NewTaskItem: typeof import('./components/NewTaskItem.vue')['default']
-    NewTaskTree: typeof import('./components/NewTaskTree.vue')['default']
     ProgressBar: typeof import('./components/ProgressBar.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
