+++ packages/ui/client/components/dashboard/TestsEntry.vue
@@ -1,15 +1,5 @@
 <script setup lang="ts">
-import { tests, testsFailed, testsSkipped, testsSuccess, testsTodo } from '~/composables/summary'
-
-const total = computed(() => tests.value.length)
-const pass = computed(() => testsSuccess.value.length)
-const failed = computed(() => testsFailed.value.length)
-const skipped = computed(() => testsSkipped.value.length)
-const todo = computed(() => testsTodo.value.length)
-// const pending = computed(() => {
-//   const t = unref(total)
-//   return t - failed.value - pass.value
-// })
+import { testStatus } from '~/composables/summary'
 </script>
 
 <template>
@@ -19,31 +9,31 @@ const todo = computed(() => testsTodo.value.length)
         Pass
       </template>
       <template #body>
-        {{ pass }}
+        {{ testStatus.testsSuccess }}
       </template>
     </DashboardEntry>
-    <DashboardEntry :class="{ 'text-red5': failed, 'op50': !failed }" data-testid="fail-entry">
+    <DashboardEntry :class="{ 'text-red5': testStatus.testsFailed, 'op50': !testStatus.testsFailed }" data-testid="fail-entry">
       <template #header>
         Fail
       </template>
       <template #body>
-        {{ failed }}
+        {{ testStatus.testsFailed }}
       </template>
     </DashboardEntry>
-    <DashboardEntry v-if="skipped" op50 data-testid="skipped-entry">
+    <DashboardEntry v-if="testStatus.testsSkipped" op50 data-testid="skipped-entry">
       <template #header>
         Skip
       </template>
       <template #body>
-        {{ skipped }}
+        {{ testStatus.testsSkipped }}
       </template>
     </DashboardEntry>
-    <DashboardEntry v-if="todo" op50 data-testid="todo-entry">
+    <DashboardEntry v-if="testStatus.testsTodo" op50 data-testid="todo-entry">
       <template #header>
         Todo
       </template>
       <template #body>
-        {{ todo }}
+        {{ testStatus.testsTodo }}
       </template>
     </DashboardEntry>
     <DashboardEntry :tail="true" data-testid="total-entry">
@@ -51,7 +41,7 @@ const todo = computed(() => testsTodo.value.length)
         Total
       </template>
       <template #body>
-        {{ total }}
+        {{ testStatus.totalTests }}
       </template>
     </DashboardEntry>
   </div>
